### 2024-07-05 f08ab10
fixes to defaults and weird tibble bug
