+++ MeDUSA/R/3_mz_subtraction.R
@@ -1,4 +1,5 @@
 # *** Subtraction -----------------------------------------------------
+#' @Description
 #' The methods of single cell sampling will often lead to the capture of cell
 #' media along with the cell itself. The media contains a significant number of
 #' metabolites, so the signal from the media should be isolated from the
