+++ MeDUSA/R/8_mzlog_analysis.R
@@ -1,6 +1,7 @@
 # *** PCA -----------------------------------------------------
 #' MZLOG-OBJ PCA
 #'
+#' @description
 #' A PCA plot is a relatively simple and well received method to compare samples
 #' and determine if they can be grouped by phenotype. The mzlog_analysis_pca
 #' function scales the log transformed data, removes any blacklisted m/z from
@@ -20,18 +21,18 @@
 #' null (only plots)
 #'
 #' @export
-mzlog_analysis_pca <- function(input_mzlog_obj,metadata, sample_blacklist = c(), plot_title = "PCA"  ) {
+mzlog_analysis_pca <- function(input_mzlog_obj,metadata, qual_col = "phenotype", sample_blacklist = c(), plot_title = "PCA"  ) {
   metadata <- local.meta_polarity_fixer(input_mzlog_obj,metadata)
   rownames(input_mzlog_obj) <- input_mzlog_obj$mz
   rownames(metadata) <- NULL
   t_mz_obj <- scale(t(dplyr::select(input_mzlog_obj,-mz)))
   t_mz_obj <- t_mz_obj[!row.names(t_mz_obj) %in% sample_blacklist ,]
   t_mz_obj <- merge(x = t_mz_obj,
-        y = subset(tibble::column_to_rownames(metadata, "sample_name"), select = "phenotype"),
+        y = subset(tibble::column_to_rownames(metadata, "sample_name"), select = qual_col),
         by = 0, all.x = TRUE)
   rownames(t_mz_obj) <- paste(t_mz_obj$Row.names, t_mz_obj$phenotype, sep = "&")
 
-  t_mz_obj <- prcomp(subset(t_mz_obj, select = -c(Row.names, phenotype)))
+  t_mz_obj <- prcomp(subset(t_mz_obj, select = -c(Row.names, get(qual_col))))
   summary(t_mz_obj)
 
   var_explained <- t_mz_obj$sdev^2/sum(t_mz_obj$sdev^2)
@@ -39,11 +40,11 @@ mzlog_analysis_pca <- function(input_mzlog_obj,metadata, sample_blacklist = c(),
   t_mz_obj$x %>%
     as.data.frame %>%
     rownames_to_column("sample_phenotype") %>%
-    separate(sample_phenotype, c("sample", "phenotype"), "&") %>%
+    separate(sample_phenotype, c("sample", get(qual_col), "&" )) %>%
 
     ggpubr::ggscatter(x = "PC1",y="PC2",
                       title = paste("PCA: ", local.mz_polarity_guesser(input_mzlog_obj)),
-                      color ="phenotype",
+                      color =get(qual_col),
                       size  = 0.5,
                       ellipse = TRUE,
                       mean.point = TRUE,
@@ -57,6 +58,7 @@ mzlog_analysis_pca <- function(input_mzlog_obj,metadata, sample_blacklist = c(),
 # *** Welch [T-Test] -----------------------------------------------------
 #' MZLOG-OBJ T-Test
 #'
+#' @description
 #' To determine if there is a statistical difference between phenotypes, the
 #' Welch T-rest is often used. The user is expected to isolate two phenotypes in
 #' the data (example provided for code to isolate) to pass into the
@@ -129,6 +131,7 @@ mzlog_analysis_welch <- function(phenoA_mz_obj, phenoB_mz_obj, adjust = 'fdr', c
 # *** Fold [change] -----------------------------------------------------
 #' MZLOG-OBJ Fold Change
 #'
+#' @description
 #' Fold change is one method of comparing two phenotypes, dividing one phenotype
 #' by the other. The user is expected to isolate two phenotypes in the data
 #' (example provided for code to isolate) to pass into the mzlog_analysis_fold
[…]
