+++ MeDUSA/DESCRIPTION
@@ -12,20 +12,23 @@ Encoding: UTF-8
 Roxygen: list(markdown = TRUE)
 RoxygenNote: 7.2.3
 Depends:
-    R (>= 4.0),
-    caret,
-    dplyr,
-    ggplot2,
-    ggpubr,
-    mzR,
-    parallel,
-    pheatmap,
-    pbapply,
-    randomForest,
-    tidyr,
-    tibble,
-    XML
+    R            (>= 4.0),
+    caret        (== 6.0-94),
+    dplyr        (>= 1.1.2),
+    ggplot2      (>= 3.4.2),
+    ggpubr       (== 0.6.0),
+    mzR          (== 2.34.0),
+    pbapply      (== 1.7-2),
+    pheatmap     (== 1.0.12),
+    pROC         (>= 1.18.5),
+    randomForest (== 4.7-1.1),
+    tidyr        (>= 1.3.0),
+    tibble       (== 3.2.1),
+    XML          (== 3.99-0.14)
 Suggests:
+    devtools (== 2.4.5),
+    parallel (== 4.3.0),
+    readxl   (== 1.4.3),
     testthat (>= 3.0.0)
 Config/testthat/edition: 3
 Config/testthat/parallel: true
