+++ .github/workflows/R.yaml
@@ -1,0 +1,24 @@
+name: Docker Image CI
+
+R_build:
+  runs-on: ubuntu-latest
+  steps:
+  - name: Checkout repo
+    uses: actions/checkout@v4
+    with:
+      fetch-depth: 0
+  - name: Setup R
+    uses: r-lib/actions/setup-r@v2
+    with:
+      use-public-rspm: true
+  - name: Install dependencies
+    uses: r-lib/actions/setup-r-dependencies@v2
+    with:
+      working-directory: MeDUSA
+      extra-packages: any::roxygen2
+      needs: roxygen2
+  - uses: r-lib/actions/check-r-package@v2
+    with:
+      working-directory: MeDUSA
+      upload-snapshots: true
+      build_args: 'c("--no-manual")'
