+++ .github/workflows/main.yaml
@@ -10,8 +10,8 @@ jobs:
     steps:
     - name: Check out the repo
       uses: actions/checkout@v4
-    - name: Set up Docker Buildx
-      uses: docker/setup-buildx-action@v3
+    #- name: Set up Docker Buildx
+    #  uses: docker/setup-buildx-action@v3
     - name: Log in to Docker Hub
       uses: docker/login-action@v3
       with:
@@ -22,13 +22,15 @@ jobs:
 #      uses: docker/metadata-action@9ec57ed1fcdbf14dcef7dfbe97b2010124a938b7
 #      with:
 #        images: thefollyllama/medusa
-    - name: Build & push the Docker image
-      uses: docker/build-push-action@v6
-      with:
-        context: .
-        tags: thefollyllama/medusa:latest
-        push: true
+  #  - name: Build & push the Docker image
+  #    uses: docker/build-push-action@v6
+  #    with:
+  #      context: .
+  #      tags: thefollyllama/medusa:latest
+  #      push: true
 #      tags: ${{ steps.meta.outputs.tags }}
 #      labels: ${{ steps.meta.outputs.labels }}
-    - name: Test & publish R package
-      run: docker build . --target medusa_test
+    - name: Build Docker
+      run: docker build  -f Dockerfile . -t thefollyllama/medusa
+    - name: Push image
+      run: docker push thefollyllama/medusa
