### 2024-07-16 cff8539
change meta$filtered_out to logical/boolean
