### 2024-08-06 #6892 @diazrenata
@laura-hetzel and @TheFollyLlama - on looking over the repo, my first question is about a general description of the software. Usually this information is in the README - e.g. what the software does, how to install it and use it, any dependencies, and examples. Could you add a page like this? It would make it much easier for reviewers and potential users of the software.

Also - is @laura-hetzel or @TheFollyLlama the author for correspondence?
