### 2024-09-03 fa5dfdf
Added target list on extraction, and removed auth in container
