### 2024-10-06 6633c58
extract on time, rather than sample
