### 2024-12-03 #6892 @ahmed-ali9
Dear @diazrenata , we would like an update of the status of the manuscript please?
Just incase this does not reach you, i have taken the liberty to also send you an email, i hope thats okay!
