+++ MeDUSA/R/6_mz_filter.R
@@ -62,6 +62,7 @@ mz_filter_missingness <- function(input_mz_obj, threshold = 0.1, msg = ""){
   if ( threshold < 1){
     threshold <- threshold * (length(input_mz_obj)-1)
   }
+  ##might be an issue with this line?
   input_mz_obj[is.na(input_mz_obj)] <- 0
 
   keep_peaks <- input_mz_obj[rowSums( dplyr::select(input_mz_obj, -mz) > 0 ) >= threshold,]
