+++ Dockerfile
@@ -20,8 +20,9 @@ RUN R -e 'remotes::install_version("dplyr",        version = "1.1.2")'   \
 &&  R -e 'remotes::install_version("Matrix",       version = "1.6.0")'   \
 &&  R -e 'remotes::install_version("ggpubr",       version = "0.6.0")'   \
 &&  R -e 'remotes::install_version("pROC",         version = "1.18.5")'  \
+&&  R -e 'remotes::install_version("duckdb",       version = "1.2.0")'    \
 &&  R -e 'remotes::install_version("readxl",       version = "1.4.3")'   \
-&&  R -e 'remotes::install_version("testthat",     version = "3.2.1.1")'
+&&  R -e 'remotes::install_version("testthat",     version = "3.2.1.1")' 
 
 FROM bioc_base
 COPY MeDUSA MeDUSA
