+++ MeDUSA/R/1_mzml_extract_centroid.R
@@ -29,39 +29,40 @@
 #' @param files \cr
 #'   String: File or directory of mzmL files
 #'   List: of mzML files (i.e. list.files(".",pattern = "qc_.*mzML"))
-#' @param cores
+#' @param parallel
 #'   Integer: Can I has multithreading? (Need parallel)
 #' @param params
 #'   List: override any subset (or all) default parameters
 #' @returns MzObj
 #' @export
-mzml_extract_magic <- function(files = getwd(), cores = 6,  params = NULL ){
+mzml_extract_magic <- function(files = getwd(), polarity = c(0,1), params = NULL ){
   start <- Sys.time()
-  params <- extract.fill_defaults(params)
   files <- extract.file_lister(files)
-
-  tryCatch({
-    if (cores > 1){
-      cl <- local.export_thread_env(2)
-      out <- parallel::parLapply(cl, c(0,1), function(x) extract.polarity_loop(polarity = x, files, cores, params))
-      names(out) <- c('neg','pos')
-      local.kill_threads(cl)
-      gc()
-      cl <- local.export_thread_env(2)
-      out <- parallel::parLapply(cl, c("neg","pos"), function(x) extract.polarity_final(x, out, files, params))
-      names(out) <- c('neg','pos')
-    } else {
-      mz_pos <- extract.polarity_loop(files,1,params)
-      print("INFO: mzml_extract_magic: Positive Complete")
-      mz_neg <- extract.polarity_loop(files,0,params)
-      out <- list(pos=mz_pos, neg=mz_neg)
-    }
-    print(paste("INFO: mzml_extract_magic SUCCESS, execution complete in:",round(as.numeric(Sys.time()-start, units="mins"),3),"minutes"))
-    return(as.data.frame(out))
-  },
-  finally={
-    local.kill_threads()
+  # Parallelizing pos/neg is too much for reasonable laptop memory
+  df <- pblapply( polarity, function(pol){
+    pol_eng <- extract.pol_english(pol)
+    params <- extract.fill_defaults(params, pol_eng)
+    print(paste0("INFO: mzml_extract_magic : Extraction of [",pol_eng,"], beginning"))
+    pbapply::pblapply(files, function(file) mzml_extract_file(file, polarity=pol, cl = NULL, magic=T, params=params))
+    print(paste0("INFO: mzml_extract_magic : Extraction of [",pol_eng,"], complete"))
+    tables <- DBI::dbGetQuery(params$dbconn, "SELECT table_name FROM duckdb_tables;")
+    query <- paste0("SELECT * FROM ", paste(tables[,] ,collapse=" UNION ALL BY NAME SELECT * FROM "), " ORDER BY mz")
+    print(paste0("INFO: mzml_extract_magic : MZ_OBJ of [",pol_eng,"], created"))
+    out <- DBI::dbGetQuery(params$dbconn, query)
+    duckdb::dbWriteTable(params$dbconn ,paste0("PREFINAL_",pol_eng) , out,append = TRUE)
+    print(paste0("INFO: mzml_extract_magic : Final Binning [",pol_eng,"], starting"))
+    #params <- extract.fill_defaults(params, pol_eng)
+    #df[pol_eng] <- DBI::dbGetQuery(params$dbconn, query)
+    gc()
+    out <- extract.mz_format(out)
+    out <- extract.binning(out, params$postbin_method, paste0('extract_magic_',pol_eng))
+    duckdb::dbWriteTable(params$dbconn ,paste0("FINAL_",pol_eng) , out,append = TRUE)
+    print(paste0("INFO: mzml_extract_magic : Final Binning [",pol_eng,"], complete"))
[…]
