### 2025-03-17 9cf5895
some db fixes
