+++ MeDUSA/R/1_mzml_extract_centroid.R
@@ -47,23 +47,29 @@ mzml_extract_magic <- function(files = getwd(), polarity = c(0,1), params = NULL
     pbapply::pblapply(files, function(file) mzml_extract_file(file, polarity=pol, cl = NULL, magic=T, params=params))
     print(paste0("INFO: mzml_extract_magic : Extraction of [",pol_eng,"], complete"))
     tables <- DBI::dbGetQuery(params$dbconn, "SELECT table_name FROM duckdb_tables;")
-    query <- paste0("SELECT * FROM ", paste(tables[,] ,collapse=" UNION ALL BY NAME SELECT * FROM "), " ORDER BY mz")
-    print(paste0("INFO: mzml_extract_magic : MZ_OBJ of [",pol_eng,"], created"))
-    out <- DBI::dbGetQuery(params$dbconn, query)
-    duckdb::dbWriteTable(params$dbconn ,paste0("PREFINAL_",pol_eng) , out,append = TRUE)
+    query <- paste0("CREATE TABLE ",paste0("PREFINAL_",pol_eng), " AS (SELECT * FROM ", paste(tables[,] ,collapse=" UNION ALL BY NAME SELECT * FROM "), " ORDER BY mz)")
+    print(paste0("INFO: mzml_extract_magic : creating initial MZ_OBJ of [",pol_eng,"]"))
+    DBI::dbGetQuery(params$dbconn, query)
     print(paste0("INFO: mzml_extract_magic : Final Binning [",pol_eng,"], starting"))
     #params <- extract.fill_defaults(params, pol_eng)
     #df[pol_eng] <- DBI::dbGetQuery(params$dbconn, query)
     gc()
+    out <- DBI::dbGetQuery(params$dbconn, paste0("select * from PREFINAL_", pol_eng))
     out <- extract.mz_format(out)
-    out <- extract.binning(out, params$postbin_method, paste0('extract_magic_',pol_eng))
+    out <- mzT_binning(out, params$postbin_method, paste0('extract_magic_',pol_eng))
     duckdb::dbWriteTable(params$dbconn ,paste0("FINAL_",pol_eng) , out,append = TRUE)
     print(paste0("INFO: mzml_extract_magic : Final Binning [",pol_eng,"], complete"))
     gc()
     out
   })
+  out <- list()
+  for( i in 1:length(polarity)) {
+    name <- extract.pol_english(polarity[i])
+    out[name] <- df[i]
+  } 
+
   print(paste("INFO: mzml_extract_magic SUCCESS, execution complete in:",round(as.numeric(Sys.time()-start, units="mins"),3),"minutes"))
-  df
+  out
 }
 
 # ***  -----------------------------------------------------
@@ -103,13 +109,13 @@ mzml_extract_file <- function(file, polarity = "",  magic = T, cl = NULL , param
   params <- extract.fill_defaults(params,pol_eng)
   name <- strsplit(basename(file),"\\.")[[1]][1]
   name <- paste(pol_eng,name,sep="_")
-  if(! name %in% DBI::dbGetQuery(params$dbconn, "show tables")[[1]] ) {
+  if( is.null(params$dbconn) || ! name %in% DBI::dbGetQuery(params$dbconn, "show tables")[[1]] ) {
     if (!is.null(cl)) {
       pbo <- pbapply::pboptions(type = "none")
       on.exit(pbapply::pboptions(pbo), add = TRUE)
     }
 
-    file <- file.path(file)
+    #file <- file.path(file)
 
     if (!file.exists(file)) {
       warning(sprintf("Cannot find file %s", file))
[…]
