### 2025-03-18 3894c8c
binning memory fixes
