+++ MeDUSA/R/1_mzml_extract_centroid.R
@@ -39,35 +39,50 @@
 mzml_extract_magic <- function(files = getwd(), polarity = c(0,1), params = NULL ){
   start <- Sys.time()
   files <- extract.file_lister(files)
-  # Parallelizing pos/neg is too much for reasonable laptop memory
-  df <- pblapply( polarity, function(pol){
-    pol_eng <- extract.pol_english(pol)
-    params <- extract.fill_defaults(params, pol_eng)
-    print(paste0("INFO: mzml_extract_magic : Extraction of [",pol_eng,"], beginning"))
-    pbapply::pblapply(files, function(file) mzml_extract_file(file, polarity=pol, cl = NULL, magic=T, params=params))
-    print(paste0("INFO: mzml_extract_magic : Extraction of [",pol_eng,"], complete"))
-    tables <- DBI::dbGetQuery(params$dbconn, "SELECT table_name FROM duckdb_tables;")
-    query <- paste0("CREATE TABLE ",paste0("PREFINAL_",pol_eng), " AS (SELECT * FROM ", paste(tables[,] ,collapse=" UNION ALL BY NAME SELECT * FROM "), " ORDER BY mz)")
-    print(paste0("INFO: mzml_extract_magic : creating initial MZ_OBJ of [",pol_eng,"]"))
-    DBI::dbGetQuery(params$dbconn, query)
-    print(paste0("INFO: mzml_extract_magic : Final Binning [",pol_eng,"], starting"))
-    #params <- extract.fill_defaults(params, pol_eng)
-    #df[pol_eng] <- DBI::dbGetQuery(params$dbconn, query)
-    gc()
-    out <- DBI::dbGetQuery(params$dbconn, paste0("select * from PREFINAL_", pol_eng))
-    out <- extract.mz_format(out)
-    out <- mzT_binning(out, params$postbin_method, paste0('extract_magic_',pol_eng))
-    duckdb::dbWriteTable(params$dbconn ,paste0("FINAL_",pol_eng) , out,append = TRUE)
-    print(paste0("INFO: mzml_extract_magic : Final Binning [",pol_eng,"], complete"))
-    gc()
-    out
-  })
+  #TODO: Fix threaded db connections
+  #if (parallel && length(polarity) > 1){
+    #cl <- local.export_thread_env(2)
+  #} else {
+    cl <- NULL
+  #}
+  #tryCatch({
+    df <- pblapply( polarity, cl=cl, function(pol){
+      pol_eng <- extract.pol_english(pol)
+      params <- extract.fill_defaults(params, pol_eng)
+      print(paste0("INFO: mzml_extract_magic : Extraction of [",pol_eng,"], beginning"))
+      pbapply::pblapply(files, function(file) mzml_extract_file(file, polarity=pol, cl = NULL, magic=T, params=params))
+      print(paste0("INFO: mzml_extract_magic : Extraction of [",pol_eng,"], complete"))
+      tables <- DBI::dbGetQuery(params$dbconn, "SELECT table_name FROM duckdb_tables;")
+      if (! paste0("PREFINAL_",pol_eng) %in% tables[,]) {
+        query <- paste0("CREATE TABLE ",paste0("PREFINAL_",pol_eng), " AS (SELECT * FROM ", tables[1,], paste(" FULL OUTER JOIN", tables[2:length(tables[,]),], collapse = " USING (mz)"), " USING (mz) ORDER BY mz)")
+        print(paste0("INFO: mzml_extract_magic : creating initial MZ_OBJ of [",pol_eng,"]"))
+        DBI::dbExecute(params$dbconn, query)
+      } else {
[…]
