### 2025-03-19 c724653
filtering qol
