### 2025-04-14 91c4393
handful of small fixes
