+++ MeDUSA/R/1_mzml_extract_centroid.R
@@ -19,24 +19,40 @@
 #' The mzml_extract_magic function performs the above mentioned alignment in one
 #' function.
 #'
-#'   Current Defaults:
-#'      "postbin_method" = max,
-#'      "tolerance" = 5e-6,
-#'      "timeSquash_method" = mean,
-#'      "missingness_threshold" = .1,
-#'      "intensity_threshold" = 1000,
-#'      "target_list" = c()
 #'
-#' @param files \cr
+#' @param files
 #'   String: File or directory of mzmL files
 #'   List: of mzML files (i.e. list.files(".",pattern = "qc_.*mzML"))
-#' @param parallel
-#'   Integer: Can I has multithreading? (Need parallel)
+#' @param polarity
+#'   0=negative, 1=positive, c(0,1)=both
+#' @param cores
+#'   Integer: Can I has multithreading? (Requires parallel and significant memory)
 #' @param params
 #'   List: override any subset (or all) default parameters
-#' @returns MzObj
+#'   Default Descriptions:
+#'     "dbconn" connection object to a database. Or NULL to ignore database features
+#'     "prebin_method" which sql_method(min, max, sum, avg) to combine intensities within the same mz bin per sample mzTs
+#'     "postbin_method" which sql_method(min, max, sum, avg) to combine intensities within the same mz bin on the final mzObj
+#'     "tolerance" tolerance for binning
+#'     "timeSquash_method" which R_method(min, max, median, mean) to comibine intensities across scans
+#'     "missingness_threshold" see mz_filter_missingness()
+#'     "intensity_threashold" see mz_filter_low_intensity()
+#'     "target_list" list of mzs to target
+#'     "target_list_tolerance" tolerance of the mz target list
+#'   Defaults Values:
+#'     "dbconn" = duckdb::dbConnect(duckdb::duckdb(paste0(local.output_dir(),'/',db_prefix,'_extract.duckdb'))),
+#'     "prebin_method" = "max",
+#'     "postbin_method" = "max",
+#'     "tolerance" = 5e-6,
+#'     "timeSquash_method" = mean,
+#'     "missingness_threshold" = 1,
+#'     "missingness_threshold" = .1,
+#'     "intensity_threshold" = 1000,
+#'     "target_list" = c(),
+#'     "target_list_tolerance" = 5e-5
+#' @returns MzObj & Database file in the output directory
 #' @export
-mzml_extract_magic <- function(files = getwd(), polarity = c(0,1), params = NULL ){
+mzml_extract_magic <- function(files = getwd(), polarity = c(0,1), cores=1, params = NULL ){
   start <- Sys.time()
   files <- extract.file_lister(files)
   #TODO: Fix threaded db connections
@@ -117,6 +133,8 @@ mzml_extract_magic <- function(files = getwd(), polarity = c(0,1), params = NULL
 #'   [0|1]: 0=Negative, 1=Positive, NULL=both
 #' @param cl \cr
 #'   parallel::threadCluster (optional)
+#' @param params
+#'   list of params, see mzml_extract_magic()
 #' @returns MzT object
 #' @export
 mzml_extract_file <- function(file, polarity = "",  magic = T, cl = NULL , params = NULL) {
@@ -217,9 +235,9 @@ mzml_extract_file <- function(file, polarity = "",  magic = T, cl = NULL , param
 #' @returns MzT object
 #' @export
 # Note missingness_threshold is very low
[…]
