### 2025-04-29 #6892 @jromanowska
Hi, @laura-hetzel :wave:
I will take care of the process from now on. I will try to find reviewers, but first I need some clarifications from you. You indicated that `@ahmed-ali9` is the corresponding author, but he did not make any contributions to the repository. Please explain how did he contribute to the work (as you can read in the [Documentation](https://joss.readthedocs.io/en/latest/submitting.html#authorship), purely organizational or financial contributions is not enough for co-authorship in JOSS paper).

Moreover, by "corresponding author", we mean the one who can handle any questions about the implementations, comments from the reviewers, etc. Usually, it's the author who also contributed most to the development of the software.

Last, but not least, could you comment on the licence choice? It's not a standard choice for scientific software.
