+++ .github/workflows/main.yaml
@@ -35,11 +35,12 @@ jobs:
   #      push: true
 #      tags: ${{ steps.meta.outputs.tags }}
 #      labels: ${{ steps.meta.outputs.labels }}
-    - name: Build Docker
-      run: |
-        docker build . -f Dockerfile -t thefollyllama/medusa:${DOCKER_TAG}-amd64  --platform linux/amd64 --provenance=FALSE
-        docker build . -f Dockerfile -t thefollyllama/medusa:${DOCKER_TAG}-arm64  --platform linux/arm64 --provenance=FALSE
-        docker manifest create thefollyllama/medusa:${DOCKER_TAG}  --amend thefollyllama/medusa:${DOCKER_TAG}-arm64  --amend thefollyllama/medusa:${DOCKER_TAG}-amd64
+    - name: Build Docker AMD
+      run: docker build . -f Dockerfile -t thefollyllama/medusa:${DOCKER_TAG}-amd64  --platform linux/amd64 --provenance=FALSE
+    - name: Build Docker ARM
+      run: docker build . -f Dockerfile -t thefollyllama/medusa:${DOCKER_TAG}-arm64  --platform linux/arm64 --provenance=FALSE
     - name: Push image
-      run: docker manifest push thefollyllama/medusa:${DOCKER_TAG}
+      run: |
+        docker manifest create thefollyllama/medusa:${DOCKER_TAG}  --amend thefollyllama/medusa:${DOCKER_TAG}-arm64  --amend thefollyllama/medusa:${DOCKER_TAG}-amd64
+        docker manifest push thefollyllama/medusa:${DOCKER_TAG}
       if: github.event_name != 'pull_request'
