+++ .github/workflows/main.yaml
@@ -10,36 +10,43 @@ on:
     branches: [ "main" ]
 
 jobs:
-  docker_build:
+  docker_build_amd:
     runs-on: ubuntu-latest
     steps:
     - name: Check out the repo
       uses: actions/checkout@v4
-    #- name: Set up Docker Buildx
-    #  uses: docker/setup-buildx-action@v3
     - name: Log in to Docker Hub
       uses: docker/login-action@v3
       with:
         username: ${{ secrets.DOCKERHUB_USER }}
         password: ${{ secrets.DOCKERHUB_TOKEN }}
-#    - name: Extract metadata (tags, labels) for Docker
-#      id: meta
-#      uses: docker/metadata-action@9ec57ed1fcdbf14dcef7dfbe97b2010124a938b7
-#      with:
-#        images: thefollyllama/medusa
-  #  - name: Build & push the Docker image
-  #    uses: docker/build-push-action@v6
-  #    with:
-  #      context: .
-  #      tags: thefollyllama/medusa:latest
-  #      push: true
-#      tags: ${{ steps.meta.outputs.tags }}
-#      labels: ${{ steps.meta.outputs.labels }}
     - name: Build Docker AMD
       run: docker build . -f Dockerfile -t thefollyllama/medusa:${DOCKER_TAG}-amd64  --platform linux/amd64 --provenance=FALSE
-    - name: Build Docker ARM
-      run: docker build . -f Dockerfile -t thefollyllama/medusa:${DOCKER_TAG}-arm64  --platform linux/arm64 --provenance=FALSE
-    - name: Push image
+    - name: Push image AMD
+      run: docker push thefollyllama/medusa:${DOCKER_TAG}-amd64
+      if: github.event_name != 'pull_request'
+
+  docker_build_arm:
+    runs-on: ubuntu-latest
+    steps:
+    - name: Check out the repo
+      uses: actions/checkout@v4
+    - name: Log in to Docker Hub
+      uses: docker/login-action@v3
+      with:
+        username: ${{ secrets.DOCKERHUB_USER }}
+        password: ${{ secrets.DOCKERHUB_TOKEN }}
+    - name: Build Docker AMD
+      run: docker build . -f Dockerfile -t thefollyllama/medusa:${DOCKER_TAG}-arm64  --platform linux/amd64 --provenance=FALSE
+    - name: Push image AMD
+      run: docker push thefollyllama/medusa:${DOCKER_TAG}-arm64
+      if: github.event_name != 'pull_request'
+
+  docker_create_manifest:
+    needs: [docker_build_amd, docker_build_arm]
+    runs-on: ubuntu-latest
+    steps:
+    - name: create_manifest
       run: |
         docker manifest create thefollyllama/medusa:${DOCKER_TAG}  --amend thefollyllama/medusa:${DOCKER_TAG}-arm64  --amend thefollyllama/medusa:${DOCKER_TAG}-amd64
         docker manifest push thefollyllama/medusa:${DOCKER_TAG}
