+++ Dockerfile
@@ -5,24 +5,24 @@ RUN R -e 'BiocManager::install("mzR")'
 COPY scripts/db_download.sh scripts/db_massager.sh ./
 RUN sh /db_download.sh; sh /db_massager.sh
 
-RUN R -e 'remotes::install_version("dplyr",        version = "1.1.2")'   \
-&&  R -e 'remotes::install_version("ggplot2",      version = "3.4.2")'   \
-&&  R -e 'remotes::install_version("tibble",       version = "3.2.1")'   \
-&&  R -e 'remotes::install_version("tidyr",        version = "1.3.0")'   \
-&&  R -e 'remotes::install_version("pbapply",      version = "1.7-2")'   \
-&&  R -e 'remotes::install_version("devtools",     version = "2.4.5")'   \
-&&  R -e 'remotes::install_version("pheatmap",     version = "1.0.12")'  \
-&&  R -e 'remotes::install_version("XML",          version = "3.99-0.14")' \
-&&  R -e 'remotes::install_version("randomForest", version = "4.7-1.1")' \
-&&  R -e 'remotes::install_version("caret",        version = "6.0-94")'  \
-&&  R -e 'remotes::install_version("caTools",      version = "1.18.2")'  \
-&&  R -e 'remotes::install_version("AICcmodavg",   version = "2.3-2")'   \
-&&  R -e 'remotes::install_version("Matrix",       version = "1.6.0")'   \
-&&  R -e 'remotes::install_version("ggpubr",       version = "0.6.0")'   \
-&&  R -e 'remotes::install_version("pROC",         version = "1.18.5")'  \
-&&  R -e 'remotes::install_version("duckdb",       version = "1.2.0")'    \
-&&  R -e 'remotes::install_version("readxl",       version = "1.4.3")'   \
-&&  R -e 'remotes::install_version("testthat",     version = "3.2.1.1")' 
+RUN R -e 'remotes::install_version("dplyr",        version = "1.1.4")' \
+ &&  R -e 'remotes::install_version("ggplot2",      version = "3.5.2")'  \
+ &&  R -e 'remotes::install_version("tibble",       version = "3.2.1")'  \
+ &&  R -e 'remotes::install_version("tidyr",        version = "1.3.1")'  \
+ &&  R -e 'remotes::install_version("pbapply",      version = "1.7-2")'  \
+ &&  R -e 'remotes::install_version("devtools",     version = "2.4.5")'  \
+ &&  R -e 'remotes::install_version("pheatmap",     version = "1.0.12")' \
+ &&  R -e 'remotes::install_version("XML",          version = "3.99-0.18")' \
+ &&  R -e 'remotes::install_version("randomForest", version = "4.7-1.2")' \
+ &&  R -e 'remotes::install_version("caret",        version = "7.0-1")'   \
+ &&  R -e 'remotes::install_version("caTools",      version = "1.18.3")'  \
+ &&  R -e 'remotes::install_version("AICcmodavg",   version = "2.3-4")'   \
+ &&  R -e 'remotes::install_version("Matrix",       version = "1.7-3")'  \
+ &&  R -e 'remotes::install_version("ggpubr",       version = "0.6.0")'  \
+ &&  R -e 'remotes::install_version("pROC",         version = "1.18.5")' \
+ &&  R -e 'remotes::install_version("duckdb",       version = "1.2.1")'  \
+ &&  R -e 'remotes::install_version("readxl",   version = "1.4.5")'  \
+ &&  R -e 'remotes::install_version("testthat", version = "3.2.3")'
 
 FROM bioc_base
 COPY MeDUSA MeDUSA
