### 2025-05-02 #6892 @laura-hetzel
Hi @jromanowska ,
First of all, very excited this is gaining traction again!
@ahmed-ali9 is heavily involved in the conceptualization of this project, and he will continue to build on it with other students. Some of the new developments are currently in progress. Also, the logic behind the functions and data processing was developed by him. I will be finishing my PhD soon (yay!) and this project will be passed on, however, Ahmed will remain the point of contact and have the overview of any updates.
The license was recommended by our IT department, so we accepted it as it complies with what we want for the project as well as what the university requires.
As for the title, direct infusion is a method of mass spectrometry, and this software was developed specifically for this mehtod. Most software in our field is developed for a different method and is therefore not useful to us, so we thought it was important to include "direct infusion" in the title. MeDUSA stands for MEtabolomics for Direct-infusion Untargeted Single-cell Analysis and this captures all of the important concepts that sets the software apart from anything else currently available.
