### 2025-05-02 #6892 @jromanowska
Thank you for the clarifications, @laura-hetzel. Wishing you luck with wrapping up your PhD - it always seems far too much to do in the last months, but you'll make it!

The conceptualization is enough to be on the author list, but I still think that since it was you who implemented that, you should be the corresponding author on this submission.

I will have to ask the other editors about the license, but I think it's fine.

Thanks for the clarification of "direct infusion" - I was wondering more about the "staring down" phrase... I'm not a native English speaker and I didn't really catch the meaning of this title. (I'm sure it's fine, I just wanted to be able to understand that 😅 )
