### 2025-05-03 #6892 @Kevin-Mattheus-Moerman
@jromanowska the license `artistic-2.0` seems okay. It is the first time I see this one but it does seem to be an OSI approved license and therefore acceptable: https://opensource.org/license/artistic-2-0.

I recommend that the authors list, link to, and name the license in their README if they don't already (and even to create a license badge for it for clarity). Also they should check whether the license text is unaltered. It should be exactly like the text from the link. I say this because this may be why GitHub lists it as `Other` at the moment.
