### 2025-05-08 #6892 @ahmed-ali9
> Thank you for the clarifications, [@laura-hetzel](https://github.com/laura-hetzel). Wishing you luck with wrapping up your PhD - it always seems far too much to do in the last months, but you'll make it!
>
> The conceptualization is enough to be on the author list, but I still think that since it was you who implemented that, you should be the corresponding author on this submission.
>
> I will have to ask the other editors about the license, but I think it's fine.
>
> Thanks for the clarification of "direct infusion" - I was wondering more about the "staring down" phrase... I'm not a native English speaker and I didn't really catch the meaning of this title. (I'm sure it's fine, I just wanted to be able to understand that 😅 )
>
> And one more thing - do you have a non-cis-male candidate for a reviewer?

Hi @jromanowska ,
I do not mind not being corresponding author for this (especially that laura needs to graduate, so avoiding delays is the name of the game now).
Please keep in mind that as she stated, she is moving on to brighter pastures after this, so she will likely not be able to respond in a timely manner to requests. Since i was involved from the beginning and i'm continuing this project as well, thats why we opted to put me as corresponding author.

Perhaps we can both be co-corresponding authors?
If its not possible, i'm happy with laura being corresponding author.

Regarding the reviewer, I can suggest the following:
l.b.zwep@lacdr.leidenuniv.nl.   ;  https://laurazwep.com/
Dr. Marina Meester ;  https://www.uu.nl/staff/MMeester
 Prof. BERNASCONI ANNA  ;https://www.deib.polimi.it/eng/people/details/711462

Thanks!
