### 2025-05-08 #6892 @ahmed-ali9
> Thank you, [@ahmed-ali9](https://github.com/ahmed-ali9). Now all is clear! 🙌 I will, then refer to both of you when writing to authors.
>
> And thank you for the reviewer suggestions!
>
> (I'd really appreciate explanation of the title, as I mentioned above)

Hi!
Staring down can be thought of as "conquering" or "dealing with". I hope this helps :)
