### 2025-05-08 #6892 @jromanowska
I've found this definition and it might be that in this light your title is not the best: https://dictionary.cambridge.org/dictionary/english/stare-down ... (I'm just trying to wrap my head around it - just thinking that it would be difficult to find a reviewer if the title is not understood well)...
