### 2025-05-12 #6892 @laura-hetzel
In an effort to not get stuck on the title, we propose changing the title to "Taming your metabolic datasets with MeDUSA". @jromanowska if you find this more suitable, I will change it in the pdf on github.
