### 2025-05-21 #6892 @waxlamp
@jromanowska, I am willing to review, but I'm a software engineer rather than a metabolomics expert.
