### 2025-06-09 c5afb02
adding more rf outputs
