### 2025-06-19 #6892 @jromanowska
👋 @sunbrn, @ByrumLab & @joannawolthuis, @florian-huber would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: [joss.readthedocs.io/en/latest/review_criteria.html](https://joss.readthedocs.io/en/latest/review_criteria.html)
