### 2025-06-26 #6892 @jromanowska
Just wanted to let you know that I'm on summer holidays entire July and won't be able to do much then, apart from assigning another reviewer, if they respond.
See you in August!
