### 2025-08-06 #6892 @egonw
ping @Mathelab @schymane @jorainer @DeniseSl22
