### 2025-08-06 #6892 @waxlamp
> are you still up to starting the review in one or two weeks time?

Hi @jromanowska -- sorry to do this but my work commitments are piling up for the months of August and September, so I'd prefer to bow out ☹️
