### 2025-08-16 #6892 @KristinaGagalova
This is the check for pre-review items. Please address accordingly

- [x] **License**: LICENSE file is present – confirm OSI-approved text
- [x] **Scholarly effort**: Sufficient commits and contributions
- [x] **Statement of need**: Present in README – could be clarified further, with an extended description of the functionality, also for new users
- [x] **Installation instructions**: Provided (Docker-based) – what version of Docker do you recommend?
- [ ] **Example usage**: Descriptions exist, but runnable examples/datasets are missing
- [ ] ~~**API documentation**: ~~
- [ ] **Community guidelines**: No CONTRIBUTING.md or contribution instructions
- [ ] **Testing**: No visible automated tests or manual validation instructions
- [x] **Authorship**: Contributors are clearly listed
- [ ] **Versioning & Release**: No tagged release – recommend creating one
