+++ MeDUSA/R/4_mz_mass_defect.R
@@ -26,6 +26,10 @@
 #'
 #' Dependencies : ggplot2, ggpubr, dplyr
 #' @return Returns an MZ-OBJ
+#' @examples
+#'   mz_mass_defect(input_mz_obj) : Run mass defect
+#'   mz_mass_defect(input_mz_obj, magicNumber1 = 0.003, magicNumber2 = 0.03)
+#'     : Run mass defect with different parameters.
 #' @export
 
 mz_mass_defect <- function(input_mz_obj, plot = TRUE, magicNumber1 = 0.00112, magicNumber2 = 0.01953) {
