### 2025-09-01 e489ddf
mostly doc updates
