### 2025-09-08 #6892 @jromanowska
@ahmed-ali9, are you still interested in submitting this software to JOSS? I'm afraid that if we wait longer, the reviewers will not have time to work on this.
