### 2025-09-12 06af8f4
pipeline syntax
