### 2025-09-15 #6892 @jromanowska
Thank you, @TheFollyLlama, for this work. Let's start the real review!
