### 2025-09-19 laura-hetzel/MeDUSA#16 @ByrumLab
Clarify installation instructions for Docker + example files
Currently, the README does not explain how to access the example files when using the Docker-based RStudio environment. The docker image /home/local directory is empty.

Suggested improvement:
- Add instructions for cloning the repository (or mounting it) inside the Docker container.
- Provide an example command (e.g., `git clone https://github.com/laura-hetzel/MeDUSA.git` in the RStudio terminal).

This suggestion is part of the JOSS review: openjournals/joss-reviews#8993
