### 2025-09-19 laura-hetzel/MeDUSA#17 @ByrumLab
Add a vignette tutorial describing modular workflows
Figure 1 is a nice overview of the capabilities of MeDUSA. A tutorial describing the functions and steps for the example data will help users quickly change or add functions to the overall workflow.

This suggestion is part of the JOSS review: openjournals/joss-reviews#8993
