### 2025-09-19 laura-hetzel/MeDUSA#18 @ByrumLab
Errors in validate.R using docker image
I am running MeDUSA through a web browser using the provided examples/mzml/validate.R script. I was able to generate the mzml/output QualityCheck figures but was unable to generate the pca and volcano plots.

pp_out <- mz_post_magic(mz_obj_p, metadata = meta, plot = TRUE)
[1] "WARN: MeDUSA::mz_post_imputation: low_noise is < 500. If [input_mz_obj] has an mz column, ensure it is labled 'mz'"
[1] "INFO: imputation success"
[1] "INFO:MeDUSA::polarity_guess: Detected Positive"
[1] "INFO:MeDUSA::polarity_guess: Detected Positive"
[1] "INFO:MeDUSA::polarity_guess: Detected Positive"
[1] "INFO: normalization success"
[1] "INFO:MeDUSA::polarity_guess: Detected Positive"
[1] "INFO: pivot_longer success"
[1] "INFO:MeDUSA::polarity_guess: Detected Positive"
[1] "INFO: mzlong_log_transform success"
Warning message:
Removed 178 rows containing non-finite outside the scale range (`stat_boxplot()`).

> mzlog_analysis_pca(mztools_filter(pp_out$mzLog,meta,"cell","type"), meta)
[1] "INFO:MeDUSA::polarity_guess: Detected Positive"
Error in local.mz_polarity_guesser(input_mz, pos_return = "pos", neg_return = "neg") :
  ERROR: MeDUSA::polarity_guesser: Could not guess Positive or Negative from colnames

This is part of the JOSS review: openjournals/joss-reviews#8993
