### 2025-09-20 818ff8c
issue16 update readme issue18 added cell data via lfs to examples
