+++ examples/mzml/validate.R
@@ -2,6 +2,7 @@ library('MeDUSA')
 
 # Assumes Docker environment
 # set the working directory to where the mzml files are located
+# Don't forget to `git lfs pull` to get the example data
 setwd("~/local/examples/mzml")
 # introduce the meta data to use for filtering
 file_meta <- "meta.xlsx"
@@ -9,7 +10,8 @@ meta <- data.frame(readxl::read_excel(file_meta, sheet = "meta", col_names = TRU
 meta$sample_name <- meta$filename
 meta <- meta[meta$sample_name %in% 
        c("sumr_blank_001", "sumr_blank_048", "sumr_blank_052", 
-         "sumr_blank_095", "sumr_qc_004", "sumr_qc_015"),]
+          "sumr_qc_004", "sumr_qc_015", "sumr_hek_079", "sumr_hek_161",
+          "sumr_hep_031", "sumr_hep_032"),]
 meta$filtered_out <- meta$filtered_out == TRUE
 
 meta <- dplyr::select(meta,-filename)
@@ -54,13 +56,14 @@ mzlog_analysis_heatmap(pp_out$mzLog[1:100,], plot_label_size = c(5,2), plot_labe
 
 ##RandomForest:Magic
 rf_data <- mztools_filter(pp_out$mzLog, meta, c("HepG2","HEK293T"))
-rf_mag <- mzlog_rf_magic(rf_data, meta, "pos")
+#rf_mag <- mzlog_rf_magic(rf_data, meta, "pos")
+
 ##RandomForest: without magic
-#rf_cor <- mzlog_rf_correlation(rf_data)
-#rf_sel <- mzlog_rf_select(rf_cor, meta)
-#rf_obj <- mzlog_rf(rf_data, meta, rfe_obj = rf_sel)
-#rf_validate <- rf_validate(rf_obj)
-#rf_permuted <- rf_permuted(rf_obj, "pos")
+rf_cor <- mzlog_rf_correlation(rf_data)
+rf_sel <- mzlog_rf_select(rf_cor, meta)
+rf_obj <- mzlog_rf(rf_data, meta, rfe_obj = rf_sel)
+rf_validate <- rf_validate(rf_obj)
+rf_permuted <- rf_permuted(rf_obj, "pos")
 
 anova <- mzlong_analysis_anova(pp_out$mzLong, meta, phenotypes=c("HepG2","HEK293T"))
 
@@ -68,48 +71,6 @@ hmdb_rf <- identify_hmdb(rf_mag$mz, c("M-H","M+Na"))
 hmdb_anova <- identify_hmdb(anova$imp_mz)
 lipids_rf <- identify_lipids(rf_mag$mz)
 lipids_rf <- identify_lipids(anova$imp_mz)
-
-
-
-
-#### Miscelaneous troubleshooting
-library('MeDUSA')
-library('duckdb')
-setwd("~/local/examples/raw/out")
-files <- list.files(getwd(), pattern="*.mzML")
-out <- mzml_extract_file(files[1], magic=T, params=list('dbconn'=NULL)) 
-mzT <- pbapply::pblapply(files, function(x) mzml_extract_file(x, polarity=0, cl = NULL, magic=T))
-mz <- mzml_extract_magic(files)
-
-method <- 'max'
-extract.binning(out)
-
-bin <- binning(out$mz, tolerance = 5e-6)
-out$mz <- bin
-cols <- as.numeric(colnames(select(out, -mz)))
-cols <- sprintf('%s("%s")', method, paste(cols ,collapse=paste0('"),',method,'("')))
-query <- paste0("SELECT mz, ",cols," FROM out GROUP BY mz" )
-
[…]
