### 2025-10-06 #8993 @ByrumLab
Hi @jromanowska, @laura-hetzel, @TheFollyLlama, @ahmed-ali9,  @KristinaGagalova. I'm not finding a workable example to test the functionality of the new software. Is there an example script somewhere? I added a few issues related to this in the repo (shown above).  Once I have a working example, I should be able to finish the review. Thanks!
