### 2025-10-06 #8993 @TheFollyLlama
Hi @ByrumLab, I added some additional sample data to (hopefully) get past the issues you encountered.
