### 2025-10-10 laura-hetzel/MeDUSA#20 @ByrumLab
mzml_extract_magic error loading qc_015 mzml file
loading in the mzml files I got an error for the sumr_qc_015.mzML file:

sumr_qc_004.mzML file size = 43.6MB
sumr_qc_015.mzML file size = 29.4MB

#Magic
mzL <- mzml_extract_magic("data")

[1] "INFO:MeDUSA::mz_filter_lowIntensity data/sumr_hek_079.mzML pos: Before Rows  : 11654"
[1] "INFO:MeDUSA::mz_filter_lowIntensity data/sumr_hek_079.mzML pos: After Rows   : 7222"
[1] "INFO:MeDUSA::mz_filter_lowIntensity data/sumr_hek_079.mzML pos: Dropped Rows : 4432"
[1] "INFO:MeDUSA::mz_filter_missingness data/sumr_hek_079.mzML pos: Before Rows  : 7222"
[1] "INFO:MeDUSA::mz_filter_missingness data/sumr_hek_079.mzML pos: After Rows   : 0"
[1] "INFO:MeDUSA::mz_filter_missingness data/sumr_hek_079.mzML pos: Dropped Rows : 7222"

[1] "INFO: Scans found in data/sumr_qc_015.mzML neg: 72"
[1] "INFO:MeDUSA::mz_filter_lowIntensity data/sumr_qc_015.mzML neg: Before Rows  : 10845"
[1] "INFO:MeDUSA::mz_filter_lowIntensity data/sumr_qc_015.mzML neg: After Rows   : 10708"
[1] "INFO:MeDUSA::mz_filter_lowIntensity data/sumr_qc_015.mzML neg: Dropped Rows : 137"
[1] "INFO:MeDUSA::mz_filter_missingness data/sumr_qc_015.mzML neg: Before Rows  : 10708"
[1] "INFO:MeDUSA::mz_filter_missingness data/sumr_qc_015.mzML neg: After Rows   : 1744"
[1] "INFO:MeDUSA::mz_filter_missingness data/sumr_qc_015.mzML neg: Dropped Rows : 8964"
[1] "INFO: mzml_extract_magic : Extraction of [neg] complete, now formatting data."
Error in checkForRemoteErrors(val) :
  one node produced an error: error reading from connection
[1] "WARN: Thread kill error: Error in defaultCluster(cl): no cluster 'cl' supplied and none is registered\n"
[1] "INFO: Killing threads. May throw odd errors."
Error in unserialize(node$con) : error reading from connection
Calls: <Anonymous> ... doTryCatch -> recvData -> recvData.SOCKnode -> unserialize
Execution halted
Error in unserialize(node$con) : error reading from connection
Calls: <Anonymous> ... doTryCatch -> recvData -> recvData.SOCKnode -> unserialize
Execution halted
Error in unserialize(node$con) : error reading from connection
Calls: <Anonymous> ... doTryCatch -> recvData -> recvData.SOCKnode -> unserialize
Execution halted
Error in unserialize(node$con) : error reading from connection
Calls: <Anonymous> ... doTryCatch -> recvData -> recvData.SOCKnode -> unserialize
Execution halted

openjournals/joss-reviews#8993
