### 2025-10-13 35b8fe2
add clarity around heatmap
