### 2025-10-20 laura-hetzel/MeDUSA#22 @KristinaGagalova
Add documentation about parallel processing and how to optimize it
MeDUSA doesn't create clusters internally; all functions are serial by default and accept cl for user-controlled parallelism. Could you please provide documentation on how to implement these steps?
Especially when we need to tailor to the user's computational resources
```
parallel
pballpy
```
This suggestion is part of the JOSS review: https://github.com/openjournals/joss-reviews/issues/8993
