### 2025-10-20 laura-hetzel/MeDUSA#23 @KristinaGagalova
Check mzT_binning()  - is the output correct (bin index)?
1. Can you please check the function?
```
> mzT_binning
function (df, method = max, log_name = "", tolerance = 5e-06)
{
    bin <- binning(df$mz, tolerance = tolerance)
    df <- suppressWarnings(aggregate(dplyr::select(df, -mz),
        list(bin), method, na.rm = T, na.action = NULL))
    names(df)[1] <- "mz"
    df[df < 0] <- 0
    df[is.na(df)] <- 0
    df
}
```
_"The intensity does not undergo further math and is simply assigned to the corresponding m/z value."_
How does that differ from [Spectra](https://www.bioconductor.org/packages/devel/bioc/vignettes/Spectra/inst/doc/Spectra.html) and why? Can you confirm that the following steps do not expect m/z instead?

2. Can you please see `log_name = ""`? It's not used in the function
3. What about the other spectral parameters, such as `ppm`, `mzFun` and optional `weighted`? I am looking at the combinePeaks [HERE](https://bioconductor.org/packages/devel/bioc/manuals/Spectra/man/Spectra.pdf)

https://github.com/openjournals/joss-reviews/issues/8993
