### 2025-10-20 laura-hetzel/MeDUSA#24 @KristinaGagalova
Analysis ANOVA
About this function here, can you please check/add the points below?
```
mzlong_analysis_anova
function (input_mzlong_obj, metadata, phenotypes, p_cutoff = 0.1,
    cores = 2)
{
    metadata <- local.meta_polarity_fixer(input_mzlong_obj, metadata)
    mzlong <- dplyr::left_join(input_mzlong_obj, metadata[c("sample_name",
        "phenotype")], by = "sample_name")
    mzlong <- mzlong[mzlong$phenotype %in% phenotypes, ]
    uniq_mz <- unique(mzlong$mz)
    cl <- local.export_thread_env(cores, environment())
    tryCatch({
        anova <- pbapply::pblapply(uniq_mz, anova.by_mz, mzlong = mzlong)
        summary <- pbapply::pblapply(seq_along(anova), anova.summary,
            anova = anova, mzs = uniq_mz)
        summary <- do.call(rbind, summary)
        tukey <- pbapply::pblapply(seq_along(anova), anova.get_p,
            anova = anova, mzs = uniq_mz)
        tukey <- do.call(rbind, tukey)
    }, finally = {
        local.kill_threads(cl)
    })
    summary <- summary[order(-summary$`Pr(>F)`, decreasing = T),
        ]
    imp_mz <- sort(summary$mz[summary$`Pr(>F)` < p_cutoff])
    list(anova = anova, summary = summary, tukey = tukey, imp_mz = imp_mz)
}
```

1. Paralellization is not used; this `cl <- local.export_thread_env(...)` is not passed to `pbapply::pblapply`. Is that intended?
2. This `order(-summary$`Pr(>F)`, decreasing = T)` is redundant. Use `order()` instead
3. Since you are performing multiple tests, it would be  more correct to add FDR correction (`p.adjust`), possibly before
`summary <- summary[order(-summary$`Pr(>F)`, decreasing = T), ]`

https://github.com/openjournals/joss-reviews/issues/8993
