### 2025-10-25 3fac267
better placed threading, and docs
