+++ MeDUSA/R/8_rf_validation.R
@@ -42,8 +42,8 @@ rf_validate <- function(rf_obj, mtry_range = c(1:200), trees = 500, mtry_seed = 
                                  "prediction" = pred_test$pred))
 
   ml <- data.frame(mtry = mtry_range)
-  cl <- local.export_thread_env(cores, environment())
   tryCatch({
+    cl <- local.export_thread_env(cores, environment())
     ml$err_rate <- pbapply::pbapply(ml, 1, rf.mtry_fit, cl = cl,
                                    data = rf_obj$train, seed = mtry_seed, trees = trees)
     best_mtry <- ml[ml$err_rate == min(ml$err_rate),]
