### 2025-10-25 d6cdec4
rf cor improvements; extract no magic steps
