+++ examples/mzml/validate.R
@@ -23,22 +23,34 @@ meta <- dplyr::select(meta,-filename)
 #load("mzL.Rdata")
 #Magic
 mzL <- mzml_extract_magic("data")
-## MzT extraction without Magic (Good for debugging)
-#files <- list.files(path=getwd("data"), pattern="*.mzML")
-#mzT <- pbapply::pblapply(files, function(x) mzml_extract_file(x, polarity=0, cl = NULL, magic=F))
-#mz <- pbapply::pblapply(mzT$pos, function(x) mzT_filtering( x, log_name = paste0(colnames(x),"pos"))
-
+mz_obj <- mzL$neg
 #save(mzL, file = 'mzL.Rdata')
 #load("mzL.Rdata")
+
+## MzT extraction without Magic (Good for debugging).=
+files <- list.files(path=paste0(getwd(),"/data"), pattern="*.mzML")
+files <- paste0("data/",files)
+mzT_raw <- pbapply::pblapply(files, function(x) mzml_extract_file(x, polarity=0, cl = NULL, magic=F))
+mzT_filt <- pbapply::pblapply(mzT_raw, function(x) mzT_filtering( x, log_name = "manual_neg"))
+mzT <- pbapply::pblapply(mzT_filt, function(x) mzT_squash_time(x))
+names <- unlist(lapply(files, function(x) strsplit(basename(x),"\\.")[[1]][1]))
+for ( x in 1:length(names)) { colnames(mzT[[x]])[2] <- paste0("neg_", names[x])}
+#Note, for smaller datasets this is less memory intense than the duckdb join. But be warned, it does not scale well.
+mz_raw <- Reduce(
+  function(x, y, ...) merge(x, y, all = TRUE, by= "mz", ...),
+    mzT
+)
+mz_obj <- mzT_binning(mz_raw)
+
      
 # Quality, filtering & post processing
-mz_obj_p <- mzL$pos
-tech_quality_p <- mz_quality_magic(mz_obj_p, meta)
-mz_obj_p <- mz_subtraction(mztools_filter(mz_obj_p, meta,"IS_Blank","type",T,T),
-                           mztools_filter(mz_obj_p, meta,"IS_Blank","type",F,T))
-mz_obj_p <- mz_mass_defect(mz_obj_p, plot = TRUE)
-mz_obj_p <- mz_filter_magic(mz_obj_p, blacklist= F)
-pp_out <- mz_post_magic(mz_obj_p, metadata = meta, plot = TRUE)
+gc()
+tech_quality <- mz_quality_magic(mz_obj, meta)
+mz_obj <- mz_subtraction(mztools_filter(mz_obj, meta,"IS_Blank","type",T,T),
+                           mztools_filter(mz_obj, meta,"IS_Blank","type",F,T))
+mz_obj <- mz_mass_defect(mz_obj, plot = TRUE)
+mz_obj <- mz_filter_magic(mz_obj, blacklist= F)
+pp_out <- mz_post_magic(mz_obj, metadata = meta, plot = TRUE)
 
 mzlog_analysis_pca(mztools_filter(pp_out$mzLog,meta,"cell","type"), meta)
 
@@ -46,13 +58,13 @@ mzlog_analysis_pca(mztools_filter(pp_out$mzLog,meta,"cell","type"), meta)
 mzlog_analysis_volcano_magic(mztools_filter(pp_out$mzLog,meta,"HepG2"),
                              mztools_filter(pp_out$mzLog,meta,"HEK293T"))
 
-welch_p <- mzlog_analysis_welch( mztools_filter(pp_out$mzLog,meta,"HepG2"),
[…]
