+++ MeDUSA/R/1_mzml_extract_centroid.R
@@ -67,8 +67,11 @@ mzml_extract_magic <- function(files = getwd(), polarity = c(0,1), params = NULL
     df <- pblapply( polarity, cl=cl, function(pol){
       pol_eng <- extract.pol_english(pol)
       params <- extract.fill_defaults(params, pol_eng)
+      if (is.null(params$dbconn)) {
+        stop("MeDUSA::mzml_extract_magic(): dbconn is currently required, do not set to NULL")
+      }
       print(paste0("INFO: mzml_extract_magic : Extraction of [",pol_eng,"], beginning"))
-      pbapply::pblapply(files, function(file) mzml_extract_file(file, polarity=pol, cl = NULL, magic=T, params=params))
+      pbapply::pblapply(files, function(file) mzml_extract_file(file, polarity=pol, cl = NULL, magic=T, params=params, return_mzobj = F))
       print(paste0("INFO: mzml_extract_magic : Extraction of [",pol_eng,"], complete"))
       tables <- DBI::dbGetQuery(params$dbconn, "SELECT table_name FROM duckdb_tables;")
       if (! paste0("PREFINAL_",pol_eng) %in% tables[,]) {
@@ -151,7 +154,7 @@ mzml_extract_magic <- function(files = getwd(), polarity = c(0,1), params = NULL
 #'          This is good for troubleshooting, it will return a list of mzT objects with limited no additional processing.
 #' @return MzT object
 #' @export
-mzml_extract_file <- function(file, polarity = "",  magic = T, cl = NULL, return_mzobj = F , params = NULL) {
+mzml_extract_file <- function(file, polarity = "",  magic = T, cl = NULL, return_mzobj = T , params = NULL) {
   pol_eng <- extract.pol_english(polarity)
   params <- extract.fill_defaults(params,pol_eng)
   name <- strsplit(basename(file),"\\.")[[1]][1]
@@ -201,8 +204,7 @@ mzml_extract_file <- function(file, polarity = "",  magic = T, cl = NULL, return
                             intensity_threshold = params$intensity_threshold,
                             log_name = name)
       print(sprintf("INFO:MeDUSA::mzml_extract_file: TimeSquashing %s",name))
-      out <- mzT_squash_time(out, timeSquash_method = params$timeSquash_method)
-      colnames(out) <- c('mz',name)
+      out <- mzT_squash_time(out, timeSquash_method = params$timeSquash_method, output_name = name)
     }
     if (class(params$dbconn) == 'duckdb_connection'){
       duckdb::dbWriteTable(params$dbconn ,name , out,append = TRUE)
@@ -214,6 +216,9 @@ mzml_extract_file <- function(file, polarity = "",  magic = T, cl = NULL, return
     }
   } else {
     print(paste0("INFO:MeDUSA::mzml_extract_file: [", name, "] found in db, skipping"))
+    if (return_mzobj) {
+      out <- DBI::dbGetQuery(params$dbconn, paste0("select * from ", name))
+    }
   }
 }
 
@@ -296,13 +301,17 @@ mzT_filtering <- function(mzT, prebin_method = 'max', missingness_threshold = 1,
 #'   [R method] : i.e. (mean, max, median)
 #' @param ignore_zeros \cr
 #'   Boolean: Should we set 0s to NA (to avoid effecting the math)
+#' @param cl \cr
+#'   parallel::threadCluster (optional).
+#' @param output_name \cr
+#'   String: Output name of the intensity column.
 #' @examples
 #'   mzT_squash_time(mzT) : Perform default time squashing
[…]
