### 2025-10-27 abb6d25
adding more data to help correlation and rf
