### 2025-10-27 21b8853
anova and rf notes
