### 2025-11-12 laura-hetzel/MeDUSA#23 @TheFollyLlama
Hello,
Regarding #3:
- ppm is handled in the parameter "tolerance" and defaults to 5e-6 (5ppm).
- mzFun is handled in the parameters prebin_method and postbin_method, they default to max; mean is also supported, and future development could validate the use of user defined binning functions.
- weighted was purposefully not considered for this application; but could be solved with user defined binning functions.
