### 2025-11-21 5cdaa1e
pca to volcano_magic and better hmdb outputs
