### 2025-11-30 laura-hetzel/MeDUSA#24 @KristinaGagalova
Hi @TheFollyLlama,
Could you please point me to the solution of 1? Thank you in advance
