### 2025-12-09 #8993 @TheFollyLlama
Hi @KristinaGagalova changes are here: https://github.com/laura-hetzel/MeDUSA/commit/3fac26756a0a3fb0e3f4f10e34752c40f0fbbabe. It's mainly reiterating my comment in the issue.
I wrote threading to be hidden within the relevant methods in the "cores" parameter. I wanted to avoid burdening the user to manage threads across their scripts. Note I've had to deprecate threading in extraction as managing threaded duckdb connections is currently a bit sad. Let me know if you need more information. Thanks!
