### 2025-12-14 9dc7f29
added data for rf, better example script
