### 2025-12-14 48c2ff3
added data for rf, better example script
