### 2026-01-05 #8993 @jromanowska
@TheFollyLlama , could you go through the tasks marked for author in the post above and get back with the DOI and version of the software?
