### 2026-01-08 #8993 @laura-hetzel
@jromanowska, Thanks for the clarification on the Zenodo issues, this is my first time archiving and I appreciate the patience! I belive everything is sorted now and we updated the pdf as well.
