### 2026-01-12 #8993 @rkurchin
👋 Hi, congrats on being nearly done! As AEiC for JOSS I will now help to process this submission for acceptance in JOSS. Below are some final checks, some of which may require your attention:
- [x] Make sure there is a GitHub tagged release that matches the version number listed here and on Zenodo
- [x] Check the .bib entry for Pang et al.: "et" doesn't need a period after, and also there appears to be a missing space between "narrowing" and "the"
- [x] in Smith et al., citation should be sentence case rather than Title Case and also "spectrometry" is misspelled, which I presume is not the case in the actual paper title 😉
- [x] Zhang and Vertes: also needs to be sentence case
- [x] in paper main text, please make sure there are spaces before parentheticals (citations, introduction of acronyms, etc.)
- [x] throughout: hyphenate "single-cell" ([compound adjective](https://www.grammarly.com/blog/parts-of-speech/compound-adjectives/))
- [x] same paragraph: third sentence should begin "One of these is..."
- [ ] there is some overall inconsistency about British vs. American English: e.g. use of "amongst" rather than "among" but spelling "visualization" with a Z rather than an S...please pick one (doesn't matter which) and run the text through a spell checker to ensure consistency
- [x] first paragraph in Statement of Need: sentence beginning "For instance" should either end after "visualization" or should have that comma replaced with a semicolon
- [x] same paragraph: hyphenate "subscription-based" (compound adjective again)
- [x] same paragraph: add comma after "must write their own scripts"
- [x] first paragraph of Description: hyphenate "user-friendly"
- [ ] throughout: "data objects" does not need to be hyphenated
- [x] second paragraph: either replace "as" with "that" or replace "any" with "the"
- [x] third paragraph: add "according" after "custom logic"
- [x] figure 2 caption: if using American English, need a comma before "which" or to change it to "that"
- [x] in "Research projects using the software": hyphenate "stem-cell" and "lab-wide"
- [x] I love your acknowledgements ❤️
