+++ docs/paper.bib
@@ -11,7 +11,7 @@
 @article{ ZV:2018,
   author = {{Zhang}, L. and {Vertes}, A.},
   journal = {Angew. Chem. Int.},
-  title = "{Single-Cell Mass Spectrometry Approaches to Explore Cellular Heterogeneity.}",
+  title = "{Single-cell mass spectrometry approaches to explore cellular heterogeneity.}",
   volume = 57,
   doi = {10.1002/anie.201709719},
   year = 2018
@@ -29,16 +29,16 @@
 @article{ SWOS:2006,
   author = {{Smith}, C.A. and {Want}, E.J. and {O'Maille}, G. and {Siuzdak}, G.},
   journal = {Anal. Chem.},
-  title = "{XCMS: Processing Mass Spectormetry Data for Metabolite Profiling Using Nonlinear Peak Alignment, Matching, and Identification}",
+  title = "{XCMS: Processing mass spectrometry data for metabolite profiling using nonlinear peak alignment, matching, and identification}",
   volume = 78,
   doi = {10.1021/ac051437y},
   year = 2006
 }
 
 @article{ P:2021,
-  author = {{Pang}, Z. et. al.},
+  author = {{Pang}, Z. et al.},
   journal = {Nucleic Acids Res.},
-  title = "{MetaboAnalyst 5.0: narrowingthe gap between raw spectra and functional insights.}",
+  title = "{MetaboAnalyst 5.0: narrowing the gap between raw spectra and functional insights.}",
   volume = 49,
   doi = {10.1093/nar/gkab382},
   year = 2021
