### 2026-02-03 cffb9a4
Update paper.md
