+++ docs/paper.md
@@ -2,7 +2,7 @@
 title: 'MeDUSA: Taming your metabolic datasets using an R-package for data processing of Direct Infusion Untarged Single-cell Analysis'
 tags:
   - R
-  - Single Cell Metabalomics
+  - Single-Cell Metabalomics
   - Mass Spectrometry
   - Metabolomics
   - Data Analytics
@@ -29,21 +29,21 @@ bibliography: paper.bib
 ---
 
 # Summary
-Over ten trillion cells are hard at work in the human body[@B:2013] and there can be significant heterogeneity amongst them affecting biological development, disease progression, and treatment response[@ZV:2018].
+Over ten trillion cells are hard at work in the human body[@B:2013] and there can be significant heterogeneity amongt them affecting biological development, disease progression, and treatment response[@ZV:2018].
 
-One technique to categorize this heterogeneity is single cell analysis using mass spectrometry. However this technique introduces new challenges. One of which is the small sample volume, which limits the possibilities for separation and thus makes the data non-compatible with traditional analysis pipelines. We introduce the R package `MeDUSA` for Metabolomic Direct-infusion Untargeted Single-cell Analysis.
+One technique to categorize this heterogeneity is single-cell analysis using mass spectrometry. However this technique introduces new challenges. One of which is the small sample volume, which limits the possibilities for separation and thus makes the data non-compatible with traditional analysis pipelines. We introduce the R package `MeDUSA` for Metabolomic Direct-infusion Untargeted Single-cell Analysis.
 
 `MeDUSA` is a start-to-finish analysis package allowing metabolomics researchers to focus on analytical content rather than R proficiency. `MeDUSA` handles the suggested workflow by Southam [@SWEJV:2017] from data extraction to filtering without a chromatogram, carrying the data through statistical analysis and identification of features for biological interpretation.
 
 # Statement of Need
-Due to the small volume of a single cell, direct infusion (DI), nano-electrospray ionization (nESI) is a highly suitable technique. However liquid chromatography mass spectrometry (LC-MS) is significantly more common than DI mass spectrometry, therefore most software is being developed for separation chromatography. For instance, XCMS offers filtering and statistical analysis with visualization, however, all of the filtering is reliant on the presence of a chromatogram [@SWOS:2006]. Similarly, MetaboAnalyst has an impressive interface that allows for metabolomic pathway analysis, but the full potential is only unlocked with a chromatogram [@P:2021]. Even paid subscription based software such as Thermo Fisher Scientific’s Compound Discoverer is designed to align and filter peaks based on a chromatogram [@CY:2024]. Furthermore, the mentioned software does not allow the user to define preprocessing methods such as centroiding or alignment. In contrast, the modularity of `MeDUSA` is built specifically for direct infusion data, and it lays the framework for method specification. The modularity also offers the user the ability to bypass functions, introduce external functions to filter, and process the data as they see fit, an option that other software does not offer. Currently, if researchers want this level of modularity and customization designed for single cell data, they must write their own scripts which requires proficiency in a programming language and a significant time allotment. Therefore the metabolomic community needs a software option that will enable complete, modular, and customizable processing of mass spectra without a chromatogram.
[…]
