### 2026-02-05 #8993 @jromanowska
@laura-hetzel @ahmed-ali9 @TheFollyLlama - could you rename the version on GitHub (or on Zenodo), so that they are the same? And also change the title on Zenodo, since you changed it here. Thanks!

I've gone through the paper, @rkurchin, and I think most of your comments were addressed, but I'll leave the ticking off to you. One small thing I've noticed is that there are still no spaces before the opening parentheses of citations in the first paragraph of "Summary".
