### 2026-02-06 72a4f09
Update paper.md
Minor editorial revisions
