### 2026-02-06 #8993 @rkurchin
Thanks! Last few things:
- [x] Still missing spaces before citations in first paragraph of Summary
- [x] Third sentence of second paragraph of summary still needs to be fixed: should begin "One of these is the small sample volume..."
- [x] still to fix: first paragraph in Statement of Need: sentence beginning "For instance" should either end after "visualization" or should have that comma replaced with a semicolon
- [x] still to fix: figure 2 caption: if using American English (which it seems like you are), need a comma before "which" or to change it to "that"
