### 2026-02-09 #8993 @laura-hetzel
@jromanowska The version issue has been resolved and new .md and .pdf files have been added to address the spacing before citations.
