### 2026-02-20 #8993 @rkurchin
👋 @openjournals/dev, seems we're getting the "file not found" thing again here when I click the DOI link (i.e. DOI resolves but PDF view is missing). Can you help out?
