### 2026-02-21 #8993 @xuanxu
Probably took some time to rebuild the pdfs page, but it is working now.
