### 2026-02-21 #8993 @arfon
Yeah, GitHub Pages builds can be slow sometimes. Looks like it is indeed fixed now.
