### 2026-02-23 #8993 @rkurchin
Great, seems fine to me too, thanks!

Congrats to @laura-hetzel and thanks to editor @jromanowska and reviewers @KristinaGagalova and @ByrumLab – JOSS couldn't run without your efforts! 🎉
