### 2026-02-23 #8993 @editorialbot
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
